refactor: move Alert, Loader to components/shared
